+++ CHANGELOG.md
@@ -4,11 +4,16 @@ All notable changes to this project will be documented in this file.
 ## [master]
 
 ### Added
+### Changed
+### Removed
+### Fixed
+
+## [1.11.1] - 2024-02-13
+
+### Added
 
 - [api] add http_request and http_request_error metrics counter.
 
-### Changed
-### Removed
 ### Fixed
 
 - [janitor] `update-idents` unables to process when some Error docs are in the index.